fix(test): keep projection setup scoped to projection support
