+++ test/test_helper.exs
@@ -1,3 +1,1 @@
-{:ok, _} = Commanded.TestSupport.ProjectionsSetup.start_link()
-
 ExUnit.start()
